requirements.txt
# WealthArena Runtime Dependencies
# Canonical runtime requirements used by scripts and Docker
# This file contains all dependencies needed to run the API and ingestion scripts

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
httpx>=0.27.0
pydantic==2.5.0
yfinance==0.2.18
pandas==2.2.2
numpy==1.26.4
prometheus-client==0.19.0
websockets==12.0

# Vector database for Knowledge Base
chromadb==0.5.4
posthog==2.4.2

# Async HTTP client
aiohttp==3.12.15

# Document processing
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
